requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
dj-database-url==2.3.0
Django==5.1.5
django-allauth==65.3.0
django-cors-headers==4.6.0
django-environ==0.11.2
django-filter==24.3
django-rest-knox==5.0.2
django-tailwind==3.8.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
idna==3.10
kombu==5.4.2
packaging==24.2
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.8.2
